Refactor/#235: BudgetUseCase 일부 로직을 AppSettingsUseCase로 이전
